+++ metadata/src/main/java/com/redhat/lightblue/metadata/translator/NonPersistedPredefinedFieldTranslatorFromJson.java
@@ -16,7 +16,7 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.crud;
+package com.redhat.lightblue.metadata.translator;
 
 import com.redhat.lightblue.metadata.EntityMetadata;
 import com.redhat.lightblue.metadata.Field;
