+++ metadata/src/main/java/com/redhat/lightblue/metadata/translator/TranslatorToJson.java
@@ -16,7 +16,7 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.crud;
+package com.redhat.lightblue.metadata.translator;
 
 import java.util.Arrays;
 import java.util.List;
